requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
idna==3.3
newsapi-python==0.2.6
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.8
